+++ src/js/util/index.js
@@ -16,4 +16,3 @@ export * from './promise';
 export * from './intersection';
 export * from './selector';
 export * from './style';
-export * from './touch';